Add clear filters button to all panels, make all viz views extend panel class. Closes #46
